+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## 0.13.0 - 2023-05-28
 ### Added
 - `kbgp_click_released`, `kbgp_user_action_released` and
   `kbgp_activate_released` for activating clicks and user actions after the